polus-unet-testing-plugin/src/requirements.txt

h5py==3.1.0
Pillow==8.1.0
filepattern==1.4.4